cd_with_leaks

+++ libft/libft.h
@@ -50,7 +50,7 @@ char				*ft_itoa(int n);
 int					ft_putchar(int i);
 void				ft_putstr(char *s, int *i);
 void				ft_putendl(char *s, int *i);
-void				ft_putnbr(int n);
+void				ft_putnbr(long int n);
 char				*ft_strmapi(char const *s, char (*f)(unsigned int, char));
 int					ft_strcmp(char *s1, char *s2);
 int					ft_scmp(const char *s1, const char *s2, char c);
